Fixing Syntax errors and resolving code review comments

+++ scripts/azureml-assets/azureml/assets/release/model_publish_utils/utils.py
@@ -11,10 +11,11 @@ from azureml.assets.util import logger
 class ModelUtils:
     """Download the Model from url at a given commit into the specified model directory."""
 
-    def __init__(self, model_url, model_commit_hash, model_dir):
+    def __init__(self, model_url: str, model_commit_hash: str, model_download_type: str, model_dir: Path):
         """Create the base object for Model."""
         self.model_url = model_url
         self.model_commit_hash = model_commit_hash
+        self.model_download_type = model_download_type
         self.model_dir = model_dir
 
     def _run(self, cmd, cwd: Path = "./") -> int:
@@ -35,12 +36,7 @@ class ModelUtils:
             logger.print(f"Successfully executed! Output: \n{result.stdout}")
         return result.returncode
 
-    def _delete_downloaded_artifacts(self) -> None:
-        """Delete the downloaded artifacts in case of failure."""
-        logger.print("Deleting the incomplete model artifacts")
-        self.model_dir.cleanup()
-
-    def download_model(self) -> bool:
+    def _download_git_model(self) -> bool:
         """Download the Model."""
         """
         Initialize LFS
@@ -49,15 +45,28 @@ class ModelUtils:
         Sets the HEAD to the commit hash.
         Deletes the .git folder in the downloaded artifacts.
         """
-        lfs_cmd = "git lfs install"
-        self._run(lfs_cmd)
-        clone_cmd = f"git clone {self.model_url} {self.model_dir.name}"
-        result = self._run(clone_cmd)
-        if result != 0:
-            self._delete_downloaded_artifacts()
+        isValidDownload = self._download_type()
+        if (isValidDownload):
+            clone_cmd = f"git clone {self.model_url} {self.model_dir}"
+            result = self._run(clone_cmd)
+            if result != 0:
+                return False
+            if self.model_commit_hash:
+                cmd = f"git reset --hard {self.model_commit_hash}"
+                self._run(cmd, cwd=self.model_dir)
+                self._run('mkdir model_dir', cwd=self.model_dir)
+                self._run('mv $(ls -I .git -I model_dir) model_dir', cwd=self.model_dir)
+
+        else:
+            logger.print('Model Download Unsupported')
             return False
-        if self.model_commit_hash:
-            cmd = f"git reset --hard {self.model_commit_hash}"
-            self._run(cmd, cwd=self.model_dir.name)
-        self._run("rm -rf .git", cwd=self.model_dir.name)
         return True
+
+    def download_model(self) -> bool:
+        """Prepare the Download Environment."""
+        if self.model_download_type == 'git':
+            download_success = self._download_git_model()
+            return download_success
+        else:
+            logger.print('Unsupported Model Download Method.')
+        return False
